requirements.txt
streamlit==1.32.0
pandas==2.2.0
Pillow==10.2.0
python-dotenv==1.0.1
openai==1.12.0
google-cloud-vision==3.5.0
pytz==2024.1
plotly==5.18.0 